Moved manifest fetching to ha-config-integrations
